refactor(listener): move listeners to one folder
